Move api/server/httpstatus to daemon/server/httpstatus
